# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  main.connectorAccountRequest:
    properties:
      label:
        type: string
      platform:
        type: string
    type: object
  main.connectorAccountView:
    properties:
      externalAccountId:
        type: string
      id:
        type: string
      label:
        type: string
      lastConnectedAt:
        type: string
      lastError:
        type: string
      lastSyncAt:
        type: string
      platform:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  main.connectorQRView:
    properties:
      accountId:
        type: string
      connected:
        type: boolean
      expiresAt:
        type: string
      lastError:
        type: string
      leaseKind:
        type: string
      platform:
        type: string
      qr:
        type: string
      queueLength:
        type: integer
      queuePosition:
        type: integer
      requestId:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      waitReason:
        type: string
    type: object
...
info:
  contact:
    name: CONXTOR project
    url: https://github.com/conxtor/groupmaster
  description: |-
    REST API for ingesting, analyzing, and presenting selected WhatsApp and Telegram group conversations.
    The public deployment serves this API below the /api reverse-proxy prefix.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/conxtor/groupmaster
  title: CONXTOR Messaging Group Intelligence API
  version: 0.1.0
paths:
...
  /connectors/accounts:
    get:
      parameters:
      - description: Administrators may include all accounts
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.connectorAccountView'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: List connector accounts
      tags:
      - connectors
    post:
      consumes:
      - application/json
      parameters:
      - description: Connector account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.connectorAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.connectorAccountView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create connector account
      tags:
      - connectors
  /connectors/accounts/{accountId}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Connector account UUID
        in: path
        name: accountId
        required: true
        type: string
      - description: Status or label update
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.connectorAccountView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update connector account
      tags:
      - connectors
  /connectors/accounts/{accountId}/qr:
    get:
      parameters:
      - description: Connector account UUID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.connectorQRView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get or start connector QR onboarding
      tags:
      - connectors
    post:
      consumes:
      - application/json
      parameters:
      - description: Connector account UUID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/main.connectorQRView'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Start connector QR onboarding
      tags:
      - connectors
...
schemes:
- http
- https
securityDefinitions:
  CookieAuth:
    description: Authenticated CONXTOR session cookie.
    in: cookie
    name: wagi_session
    type: apiKey
swagger: "2.0"